% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_bonn.R
\docType{data}
\name{uni_bonn}
\alias{uni_bonn}
\title{Colors of the University of Bonn}
\format{
An object of class \code{character} of length 3.
}
\source{
Color definitions are based on
\url{https://www.uni-bonn.de/de/universitaet/medien-universitaet/medien-presse-kommunikation/medien-corporate-design/ubo_manual_extern_6_2022.pdf}.
}
\usage{
uni_bonn
}
\value{
A named vector of colors (HEX/HTML codes of type character).
}
\description{
\code{uni_bonn} provides the main color and the two spot colors of the \href{https://www.uni-bonn.de/de}{University of Bonn}, Germany.
}
\details{
Additional to the full-tone color (100\%) the colors can be used with 75\%, 50\% and 25\%.

For digital use the main color \code{uniblau} is defined as \code{RGB 0/78/159} or \code{HEX #004e9f}. For print media the color definitions are \code{CMYK 100/70/0/0} or \code{Pantone 286}.

The spot color \code{unigelb} is defined as
\code{RGB 252/186/0} or \code{HEX #fcba00} for digital use
and \code{CMYK 0/30/100/0} or \code{Pantone 1235} for print media.

The second spot color \code{unigrau} is defined as
\code{RGB 144/144/133} or \code{HEX #909085} for digital use
and \code{CMYK 0/0/15/55} or \code{Pantone 416} for print media.
}
\examples{
uni_bonn
seecol(uni_bonn, main = "Colors of the University of Bonn") #view color palette
demopal(uni_bonn, type = 1,
        main = "Colors of the University of Bonn")


}
\seealso{
\code{\link{seecol}} for viewing and comparing color palettes;
\code{\link{usecol}} for using color palettes;
\code{\link{simcol}} for finding similar colors;
\code{\link{newpal}} for defining new color palettes;
\code{\link{grepal}} for finding named colors.
}
\author{
\strong{unikn}, 2023-06-03.
}
\concept{contributed color palettes}
\keyword{datasets}
